Addded path option

+++ src/main/bash/commands/besman-attest.sh
@@ -2,19 +2,36 @@
 
 function __bes_attest {
 	local environment_name env_repo environment_name version_id env_config
-	file_name=$1
-        
-        export COSIGN_PASSWORD=$(openssl rand -base64 32)
-	export COSIGN_KEY_LOCATION=$(pwd)
-        
-	cosign version 2>&1>/dev/null
+	opts=$1
+        args=$2
+
+        [[ "${opts[0]}" != "--file" ]] && [[ "${opts[0]}" != "--path" ]] && __besman_echo_red "Incorrect options passed." && __bes_help_"$command" && return 1
+
+	[[ ${#opts[@]} -eq 1 ]] && [[ "${opts[0]}" == "--file" ]] && filename="${args[1]}"
+	[[ ${#opts[@]} -eq 1 ]] && [[ "${opts[0]}" == "--path" ]] &&  __besman_echo_red "--file argument is must. Passed only --path" && __bes_help_"$command" && return 1
+	[[ ${#opts[@]} -eq 2 ]] && [[ "${opts[0]}" == "--path" ]] && filepath="${args[1]}" && filename="${args[2]}"
+	[[ ${#opts[@]} -eq 2 ]] && [[ "${opts[1]}" == "--path" ]] && filepath="${args[2]}" && filename="${args[1]}"
+
+	[[ ! -z $filepath ]] && [[ ! -d $filepath ]] && __besman_echo_red "filepath $filepath not found. Exit" && return 1
+        [[ ! -z $filepath ]] && [[ ! -f $filepath/$filename ]] && __besman_echo_red "file $filename not found at $filepath. Exit" && return 1
+        [[ -z $filepath ]] && [[ ! -f $filename ]] && __besman_echo_red "file $filename not found at $filepath. Exit" && return 1
+
+	if [ ! -z $filepath ];then
+	  wd=$(pwd)
+	  cd $filepath
+	fi
+
+	export COSIGN_PASSWORD=$(openssl rand -base64 32)
+        export COSIGN_KEY_LOCATION=$(pwd)
+
+        cosign version 2>&1>/dev/null
 
 	if [ xx"$?" != xx"0" ];then
 	  # install COSIGN
           LATEST_VERSION=$(curl https://api.github.com/repos/sigstore/cosign/releases/latest | grep tag_name | cut -d : -f2 | tr -d "v\", ")
           curl -O -L "https://github.com/sigstore/cosign/releases/latest/download/cosign_${LATEST_VERSION}_amd64.deb"
           sudo dpkg -i cosign_${LATEST_VERSION}_amd64.deb
-	  rm -rf cosign_${LATEST_VERSION}_amd64.deb
+	  [[ -f  cosign_${LATEST_VERSION}_amd64.deb ]] && rm -rf cosign_${LATEST_VERSION}_amd64.deb
         fi
 
 	if [ ! -f cosign.key ];then
@@ -22,18 +39,22 @@ function __bes_attest {
 	fi
 
          # Generate a predicate file
-        create_predicate $file_name
+        create_predicate $filename
 
-	if [ -f $file_name ];then
-            cosign sign-blob --yes --key cosign.key --bundle $file_name.bundle $file_name > $file_name.sig
-	    cosign attest $file_name --yes --key cosign.key --bundle $file_name.attest.bundle --predicate $file_name.predicate.json > $file_name.attest.sig
+	if [ -f $filename ];then
+            cosign sign-blob --yes --key cosign.key --bundle $filename.bundle $filename > $filename.sig
+	    cosign attest $filename --yes --key cosign.key --bundle $filename.attest.bundle --predicate $filename.predicate.json > $filename.attest.sig
 	else
-             __besman_echo_red "file $file_name not found."
+             __besman_echo_red "file $filename not found."
 	     return 1
 	fi
 
 	#upload attestation files
-	upload_attested $file_name
+	upload_attested $filename
+
+	if [ ! -z $filepath ];then
+	  cd $wd
+	fi
 }
 
 function create_predicate {
@@ -79,9 +100,10 @@ function upload_attested {
 	  git push origin
 	else
 	  echo ""
-          echo "Present directory is not git controlled."
-	  echo "Please upload the following files to the OSAR directory manually."
-	  echo "     cosign.pub $sigfile $predicatefile $file.attest.sig $file.attest.sig $file.bundle $file.attest.bundle"
+	  [[ ! -z $filepath ]] && __besman_echo_yellow "Filepath provided is not git controlled."
+	  [[ -z $filepath ]] && __besman_echo_yellow "Present working directory is not git controlled."
+	  __besman_echo_yellow "Please upload the following files to the OSAR directory manually."
+	  __besman_echo_yellow "     cosign.pub $sigfile $predicatefile $file.attest.sig $file.attest.sig $file.bundle $file.attest.bundle"
 	  echo ""
 	fi
 }
